+++ src/front.ml
@@ -120,11 +120,9 @@ let () =
 (* ------------------------------------------------------------------------- *)
 
 (* If [--strategy simplified] has been selected, then check that the [error]
-   token is used only at the end of productions (2021/10/31). Also, check
-   that the $syntaxerror keyword is not used. *)
+   token is used only at the end of productions (2021/10/31). *)
 
 let grammar =
-  CheckErrorTokenUsage.check_grammar grammar;
   CheckErrorTokenUsage.filter_grammar grammar
 
 (* ------------------------------------------------------------------------- *)